requirements.txt
python-dotenv==0.10.1
python-telegram-bot==11.1.0
redis==3.5.3
requests==2.21.0
vk-api==11.9.0
